+++ src/components/standard/muestrareservaciones.css
@@ -1,0 +1,96 @@
+
+.ticket-content {
+  padding: 40px;
+  text-align: center;
+  color: #090909;
+}
+
+.date {
+  color: #555;
+  margin-bottom: 30px;
+}
+
+.ticket-card {
+  margin: 0 auto;
+  background-color: white;
+  padding: 30px;
+  border-radius: 20px;
+  box-shadow: 0 20px 40px rgba(0, 0, 0, 0.3);
+  max-width: 900px;
+}
+
+.ticket-row {
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 20px;
+  flex-wrap: wrap;
+}
+
+.ticket-field {
+  flex: 1;
+  min-width: 200px;
+  margin: 10px;
+}
+
+.ticket-field label {
+  display: block;
+  color: #090909;
+  font-size: 14px;
+  margin-bottom: 5px;
+}
+
+.value {
+  background-color: #e5e5e5;
+  padding: 8px 14px;
+  border-radius: 8px;
+  font-weight: 500;
+}
+
+.status {
+  padding: 8px 14px;
+  border-radius: 8px;
+  font-weight: bold;
+}
+
+.status.accepted {
+  color: green;
+  background-color: #e6fbe9;
+  border: 1px solid #9be3a5;
+}
+
+.companions ul {
+  list-style: disc inside;
+  text-align: left;
+  margin-top: 10px;
+  background-color: white;
+  border: 1px solid #ccc;
+  padding: 10px;
+  border-radius: 6px;
+}
+
+.ticket-buttons {
+  margin-top: 30px;
+  display: flex;
+  justify-content: center;
+  gap: 20px;
+}
+
+.btn-view,
+.btn-exit {
+  padding: 10px 20px;
+  border-radius: 20px;
+  font-weight: bold;
+  cursor: pointer;
+}
+
+.btn-view {
+  background-color: #0c44cc;
+  color: white;
+  border: none;
+}
+
+.btn-exit {
+  background-color: white;
+  color: #0c44cc;
+  border: 1px solid #999;
+}
